+++ e2e/test/mocha-javascript/src/IsEven.js
@@ -1,7 +1,7 @@
 
 module.exports.isEven = function(number) {
   // Note: Implemented with a case switch statement, in order to reproduce this issue:
-  // https://github.com/stryker-mutator/stryker/issues/2469#issuecomment-690303849
+  // https://github.com/stryker-mutator/stryker-js/issues/2469#issuecomment-690303849
   let mod2 = number % 2; 
   let isEven;
   switch(mod2) {