+++ packages/instrumenter/src/transformers/babel-transformer.ts
@@ -12,7 +12,11 @@ import { AstFormat } from '../syntax';
 
 import { AstTransformer } from '.';
 
-export const transformBabel: AstTransformer<AstFormat.JS | AstFormat.TS> = ({ root, originFileName, rawContent }, mutantCollector, { options }) => {
+export const transformBabel: AstTransformer<AstFormat.JS | AstFormat.TS> = (
+  { root, originFileName, rawContent, offset },
+  mutantCollector,
+  { options }
+) => {
   // Wrap the AST in a `new File`, so `nodePath.buildCodeFrameError` works
   // https://github.com/babel/babel/issues/11889
   const file = new File({ filename: originFileName }, { code: rawContent, ast: root });
@@ -35,7 +39,7 @@ export const transformBabel: AstTransformer<AstFormat.JS | AstFormat.TS> = ({ ro
         )
           return;
         mutate(path, options).forEach((mutant) => {
-          mutantCollector.add(originFileName, mutant);
+          mutantCollector.add(originFileName, mutant, offset?.position, offset?.line);
         });
       }
     },
@@ -48,6 +52,17 @@ export const transformBabel: AstTransformer<AstFormat.JS | AstFormat.TS> = ({ ro
     },
   });
   if (mutantCollector.hasPlacedMutants(originFileName)) {
-    root.program.body.unshift(...instrumentationBabelHeader);
+    // Be sure to leave comments like `// @flow` in.
+    let header = instrumentationBabelHeader;
+    if (Array.isArray(root.program.body[0]?.leadingComments)) {
+      header = [
+        {
+          ...instrumentationBabelHeader[0],
+          leadingComments: root.program.body[0]?.leadingComments,
+        },
+        ...instrumentationBabelHeader.slice(1),
+      ];
+    }
+    root.program.body.unshift(...header);
   }
 };